refactor(v4): consume the core deepmerge in Track
`AbstractTrack` and `TrackContext` import it from `src/utils`, and
`Track/payload.ts` is deleted. The behaviour they get is the same one
they had, plus the prototype hardening the copy did not have — their
payloads are JSON parsed out of a `<script>` ref, so a `__proto__` key in
that JSON reached the object handed to `dataLayer.push()`.

The packed consumer covers the new subpath.
